Changes that I've wanted to make for awhile
Mainly it's cleaning up config and updating the related classes to reflect the changes.
The conditional recipe factory uses a switch statement (like it always should have) so that's a weight off my back
Some formatting changes but that's purely a visual thing code is not changed
Should also mention the comment in ConfigTFCM "Might wanna do a prettier file name. This will "reset" config for users". The commit doesn't include the nicer file name as frankly I don't wanna be the cause of any complaints however minor.